+++ tests/testthat/test-printing.R
@@ -11,7 +11,7 @@ test_that("grouped show method works", {
 
     dg <- group_by(d, am)
     out <- capture.output(dg)
-    expect_identical(out[1], "DataFrame with 32 rows and 11 columns")
+    expect_identical(out[1], "GroupedDataFrame with 32 rows and 11 columns")
     expect_identical(out[2], "Groups:  am ")
     expect_identical(substr(out[3], 1, 27), "                        mpg")
 })